updated telemetry component
