+++ frontend/src/components/Battery.js
@@ -1,4 +1,5 @@
-import styled from 'styled-components';
+import React from "react";
+import styled from "styled-components";
 const BatteryWrap = styled.div `
 text-align: center;
 display:grid;
